Include bind parameters in `debug_sql!`
I've implemented this as a separate AST pass, rather than the
`CollectBinds` pass of the `Debug` backend. This is because we will
ultimately want to use this same code for logging, which won't have the
`QueryFragment<Debug>` constraint. Ultimately I think I want to get rid
of the `Debug` backend entirely once logging is implemented.

I had originally implemented this by adding two additional lifetime
parameters onto `AstPass`. However, I really disliked how opaque and
random these two extra parameters felt, and it made the type harder to
document (and use). Shortening an invariant lifetime like this is
actually one of the documented use cases that `mem::transmute` is useful
for. I don't like introducing unsafe code in general, but I felt that
the tradeoff was worth it in this case.